requirements.txt
Django==5.0.*
git+https://github.com/AS207960/python-utils@da5df1c1b526b678891d19a2dbf77e299e50ede0
sentry-sdk
psycopg2-binary
gunicorn
django-crispy-forms
crispy-bootstrap5
django-xff
djangorestframework
django-cockroachdb==5.0.*
retry
requests
django-phonenumber-field[phonenumbers]
markdown2
html2text
lxml
lxml_html_clean
celery
cryptography
crispy-bootstrap4
beautifulsoup4
django-storages
boto3==1.35.99
stripe>=5.0.0
pika
PGPy